requirements_beta.txt
typer[all]>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.1
rich>=13.0.0
requests>=2.31.0
urllib3==1.26.5
django==2.1.7
django-cors-headers==3.5.0